feat: renaming and adding structure
