+++ src/getExtension/getExtension.ts
@@ -1,33 +1,4 @@
-import { DidPublicKey, IEncryptedMessage } from '@kiltprotocol/sdk-js'
-
-type This = typeof globalThis
-export interface PubSubSession {
-  listen: (
-    callback: (message: IEncryptedMessage) => Promise<void>
-  ) => Promise<void>
-  close: () => Promise<void>
-  send: (message: IEncryptedMessage) => Promise<void>
-  encryptionKeyId: DidPublicKey['id']
-  encryptedChallenge: string
-  nonce: string
-}
-
-export interface InjectedWindowProvider {
-  signWithDid: (
-    plaintext: string
-  ) => Promise<{ signature: string; didKeyUri: string }>
-  startSession: (
-    dAppName: string,
-    dAppEncryptionKeyId: DidPublicKey['id'],
-    challenge: string
-  ) => Promise<PubSubSession>
-  specVersion: '0.1'
-  version: string
-}
-
-interface ApiWindow extends This {
-  kilt: Record<string, InjectedWindowProvider>
-}
+import { ApiWindow, InjectedWindowProvider } from '../types/types'
 
 const apiWindow = window as Window & ApiWindow
 
@@ -60,7 +31,7 @@ function getWindowExtensions(): Promise<[InjectedWindowProvider][]> {
   )
 }
 
-export function injectExtension(): Promise<InjectedWindowProvider[]> {
+export function getExtensions(): Promise<InjectedWindowProvider[]> {
   apiWindow.kilt = apiWindow.kilt || {}
 
   return documentReadyPromise(
