+++ src/index.ts
@@ -1,7 +1,4 @@
-import {
-  injectExtension,
-  InjectedWindowProvider,
-  PubSubSession,
-} from './extension'
-export { injectExtension }
+import { getExtensions } from './getExtension/getExtension'
+import { InjectedWindowProvider, PubSubSession } from './types/types'
+export { getExtensions }
 export type { InjectedWindowProvider, PubSubSession }
